Reimplement `Notes` abstract class
