+++ maisim/maisim.Game/Component/Gameplay/Notes/SlidePath.cs
@@ -1,10 +1,31 @@
-﻿namespace maisim.Game.Component.Gameplay.Notes
+﻿using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Textures;
+using osuTK;
+
+namespace maisim.Game.Component.Gameplay.Notes
 {
     public class SlidePath : Notes
     {
-        public SlidePath(string textureName = "Notes/SlidePath.png") : base(textureName)
+        [BackgroundDependencyLoader]
+        private void load(TextureStore textureStore)
         {
-
+            InternalChild = new Container
+            {
+                Anchor = Anchor.Centre,
+                Origin = Anchor.Centre,
+                Size = new Vector2(50),
+                Child = new Sprite
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    FillMode = FillMode.Fill,
+                    Texture = textureStore.Get("Notes/SlidePath.png")
+                }
+            };
         }
     }
 }
